Format code

+++ src/DappTutorial.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.6;
 contract DappTutorial {
     receive() external payable {}
 
-    function withdraw(uint password) public {
-        require (password == 42, "Access Denied!");
+    function withdraw(uint256 password) public {
+        require(password == 42, "Access Denied!");
         payable(msg.sender).transfer(address(this).balance);
     }
 }
